+++ yast/plugins/authentication/__init__.py
@@ -2,6 +2,8 @@ from yast.applications import Yast
 
 from .middlewares import AuthenticationMiddleware
 
+__name__ = "authentication"
+
 
 def plugin_init(app: Yast, config: dict = {}):
     mw_cfg = config.pop("middleware", None)