+++ yast/plugins/database/__init__.py
@@ -1,3 +1,9 @@
+"""
+title: DatabaseMiddleware
+module: DatabaseMiddleware
+description:
+    数据库组建，可用于支持各种类型的数据库，需通过 plugin_init 启用
+"""
 __all__ = [
     "compile",
     "DatabaseBackend",
@@ -6,10 +12,18 @@ __all__ = [
     "get_database_backend",
 ]
 
+__name__ = "database"
+
+from yast.applications import Yast
 from yast.datastructures import DatabaseURL
 
 from .drivers.base import DatabaseBackend, DatabaseSession, DatabaseTransaction, compile
+from .drivers.mysql import MysqlBackend  # noqa
 from .drivers.postgres import PostgresBackend  # noqa
+
+
+def plugin_init(app: Yast, config: dict = {}):
+    print("plugin init")
 
 
 def get_database_backend(database_url: DatabaseURL, *args, **kwargs) -> DatabaseBackend: