+++ yast/plugins/schema/__init__.py
@@ -1,2 +1,5 @@
+__name__ = "schema"
+
+
 def plugin_init(app, config: dict):
     print("schema plugin init")  # pragma: no cover