+++ yast/plugins/session/__init__.py
@@ -2,6 +2,8 @@ from yast.applications import Yast
 
 from .middlewares import SessionMiddleware
 
+__name__ = "session"
+
 
 def plugin_init(app: Yast, config={}) -> None:
     if "secret_key" not in config: