+++ yast/plugins/template/__init__.py
@@ -3,6 +3,8 @@ import typing
 
 from yast.applications import Yast
 
+__name__ = "template"
+
 
 def plugin_init(app: Yast, config: dict = {}) -> None:
     def load_template_env(template_directory: str = None) -> typing.Any: