+++ yast/utils.py
@@ -1,0 +1,2 @@
+def get_modules(module_name: str):
+    pass