+++ changelog.md
@@ -1,5 +1,6 @@
 # v0.2.5
 1. Opt: Add ImmutableMultiDict, Auth: Request -> HttpConnection, graphql: executor_class
+2. database support: mysql
 
 # v0.2.4
 1. support template: default jinja2